telecom_kg/requirements.txt
aiohttp==3.8.5
pyyaml==6.0
tiktoken==0.5.1
neo4j==5.13.0 